More Node JS approach and modularization
